Added parsing to different communication types

+++ src/models/blockingp2pcommunication.cpp
@@ -2,7 +2,7 @@
 
 BlockingP2PCommunication::BlockingP2PCommunication(const otf2::definition::location &sender,
                                                    const otf2::definition::location &receiver,
-                                                   const uint32_t msgTag, const uint32_t msgLength,
+                                                   const uint32_t msgTag, const uint64_t msgLength,
                                                    const types::communicator &communicator,
                                                    const otf2::chrono::duration &sendTime,
                                                    const otf2::chrono::duration &receiveTime) :
